requirements.txt
feedparser==5.2.1
pkg-resources==0.0.0
python-dateutil==2.4.2
pytz==2015.6
requests==2.2.1
six==1.10.0
